requirements.txt
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
black==24.2.0
bleach==6.2.0
cffi==1.17.1
click==8.1.8
constantly==23.10.4
coverage==7.4.3
cryptography==45.0.2
daphne==4.2.0
decorator==5.2.1
dj-database-url==2.3.0
Django==5.2
django-cors-headers==4.7.0
django-debug-toolbar==4.3.0
django-environ==0.12.0
django-filter==25.1
django-ratelimit==4.1.0
django-redis==5.4.0
django-taggit==6.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-yasg==1.21.10
executing==2.2.0
flake8==7.0.0
gunicorn==23.0.0
uvicorn[standard]==0.27.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
iniconfig==2.1.0
ipython==8.22.2
isort==5.13.2
jedi==0.19.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mypy_extensions==1.1.0
numpy==2.2.5
packaging==25.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.7
pluggy==1.5.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.11.1
pycparser==2.22
pyflakes==3.2.0
Pygments==2.19.1
PyJWT==2.9.0
pyOpenSSL==25.1.0
pytest==8.0.2
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-magic==0.4.27
pytz==2025.2
PyYAML==6.0.2
redis==6.0.0
referencing==0.36.2
rpds-py==0.24.0
service-identity==24.2.0
setuptools==80.8.0
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
time-machine==2.16.0
traitlets==5.14.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
uritemplate==4.1.1
wcwidth==0.2.13
webencodings==0.5.1
whitenoise==6.9.0
zope.interface==7.2
